root: releasing version v5.0.0
Since Bruce and Yee Soong provided a lot of great guidances in
contract drafting, it's better to re-write everything from scratch
to match general public target audience instead of the lawyers.
Moreover, with the help from Claude.AI, this re-write is essential.
Hence, let's do it.

This patch releases version v5.0.0 with a complete license re-write
to match the latest regulatory and language requirements.

Signed-off-by: (Holloway) Chew, Kean Ho <me@hollowaykeanho.com>